moved from admin view to forum view
