+++ aa_forum/templates/aa_forum/modals/forum/board/topic-change-lock-state.html
@@ -33,7 +33,7 @@
                     {% translate "Cancel" %}
                 </button>
 
-                <a id="modal-button-confirm" class="btn btn-danger btn-sm" role="button" href="{% url 'aa_forum:admin_topic_change_lock_state' topic.pk %}">
+                <a id="modal-button-confirm" class="btn btn-danger btn-sm" role="button" href="{% url 'aa_forum:forum_topic_change_lock_state' topic.pk %}">
                     {% if topic.is_locked %}
                         <i class="fas fa-unlock"></i>
                         {% translate 'Unlock/Re-Open Topic' %}
