+++ aa_forum/templates/aa_forum/modals/forum/board/topic-delete.html
@@ -28,7 +28,7 @@
                     {% translate "Cancel" %}
                 </button>
 
-                <a id="modal-button-confirm" class="btn btn-danger btn-sm" role="button" href="{% url 'aa_forum:admin_topic_delete' topic.pk %}">
+                <a id="modal-button-confirm" class="btn btn-danger btn-sm" role="button" href="{% url 'aa_forum:forum_topic_delete' topic.pk %}">
                     <i class="far fa-trash-alt"></i>
                     {% translate "Delete" %}
                 </a>
