+++ aa_forum/urls.py
@@ -60,18 +60,18 @@ urlpatterns = [
     ),
     url(
         r"^topic/(?P<topic_id>[0-9]+)/change-lock-state/$",
-        admin.topic_change_lock_state,
-        name="admin_topic_change_lock_state",
+        forum.topic_change_lock_state,
+        name="forum_topic_change_lock_state",
     ),
     url(
         r"^topic/(?P<topic_id>[0-9]+)/change-sticky-state/$",
-        admin.topic_change_sticky_state,
-        name="admin_topic_change_sticky_state",
+        forum.topic_change_sticky_state,
+        name="forum_topic_change_sticky_state",
     ),
     url(
         r"^topic/(?P<topic_id>[0-9]+)/delete/$",
-        admin.topic_delete,
-        name="admin_topic_delete",
+        forum.topic_delete,
+        name="forum_topic_delete",
     ),
     # Forum URLs (This needs to be the last block!)
     url(
