+++ aa_forum/views/admin.py
@@ -12,7 +12,7 @@ from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
 
 from aa_forum.forms import EditBoardForm, EditCategoryForm
-from aa_forum.models import Boards, Categories, Topics
+from aa_forum.models import Boards, Categories
 
 
 @login_required
@@ -305,101 +305,3 @@ def ajax_board_order(request: WSGIRequest) -> JsonResponse:
         data.append({"success": True})
 
     return JsonResponse(data, safe=False)
-
-
-@login_required
-@permission_required("aa_forum.manage_forum")
-def topic_change_lock_state(
-    request: WSGIRequest, topic_id: int
-) -> HttpResponseRedirect:
-    """
-    Change the lock state of the given topic
-    :param request:
-    :type request:
-    :param topic_id:
-    :type topic_id:
-    :return:
-    :rtype:
-    """
-
-    topic = Topics.objects.get(pk=topic_id)
-
-    if topic.is_locked:
-        topic.is_locked = False
-
-        messages.success(
-            request,
-            mark_safe(_("<h4>Success!</h4><p>Topic has been unlocked/re-opened.</p>")),
-        )
-    else:
-        topic.is_locked = True
-
-        messages.success(
-            request,
-            mark_safe(_("<h4>Success!</h4><p>Topic has been locked/closed.</p>")),
-        )
-
-    topic.save()
-
-    return redirect("aa_forum:forum_board", topic.board.category.slug, topic.board.slug)
-
-
-@login_required
-@permission_required("aa_forum.manage_forum")
-def topic_change_sticky_state(
-    request: WSGIRequest, topic_id: int
-) -> HttpResponseRedirect:
-    """
-    Change the sticky state of the given topic
-    :param request:
-    :type request:
-    :param topic_id:
-    :type topic_id:
-    :return:
-    :rtype:
-    """
-
-    topic = Topics.objects.get(pk=topic_id)
-
-    if topic.is_sticky:
-        topic.is_sticky = False
-
-        messages.success(
-            request,
-            mark_safe(_('<h4>Success!</h4><p>Topic is no longer "Sticky".</p>')),
-        )
-    else:
-        topic.is_sticky = True
-
-        messages.success(
-            request,
-            mark_safe(_('<h4>Success!</h4><p>Topic is now "Sticky".</p>')),
-        )
-
-    topic.save()
-
-    return redirect("aa_forum:forum_board", topic.board.category.slug, topic.board.slug)
-
-
-@login_required
-@permission_required("aa_forum.manage_forum")
-def topic_delete(request: WSGIRequest, topic_id: int) -> HttpResponseRedirect:
-    """
-    Delete a given topic
-    :param request:
-    :type request:
-    :param topic_id:
-    :type topic_id:
-    """
-
-    topic = Topics.objects.get(pk=topic_id)
-    board = topic.board
-
-    topic.delete()
-
-    messages.success(
-        request,
-        mark_safe(_("<h4>Success!</h4><p>Topic removed.</p>")),
-    )
-
-    return redirect("aa_forum:forum_board", board.category.slug, board.slug)
